requirements.txt
# requirements.txt
pymongo==4.10.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.8
pydantic>=2.0.0
openai>=1.12.0
tavily-python==0.3.7
black==23.12.1
cohere